Create README.md
Hello! My name is Exequiel Robles Vega and I am a UX/UI designer, this means that I specialize in creating products of greater value for users, increasing their efficiency.

Throughout my life I have dedicated myself to the world of design, working as an industrial, graphic and fashion designer. In my quest to train as a professional, I entered the world of UX/UI design.

I aspire to join a company that is interested in innovating and that challenged me to grow as a professional. I also value the possibility of keep learning through courses and programs.

¡Thank you very much for your interest and I hope we can stay in touch!